KSN v0.3
I lied, looks like I chose to add a delta-V calculator tonight as well. At a later point I might add a system that imports the planet configurations from a JSON so that it's more accessible to edit. I may also include the option to save the orbital characteristics to a text file.